// frontend/src/app/api/model/leaderboardEntry.ts
/**
 * Yonna Akademia API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface LeaderboardEntry { 
    rank: number;
    user_id: number;
    username: string;
    total_xp: number;
}